Is Subsequence - Leetcode
Problem:

Given two strings s and t, return true if s is a subsequence of t, or false otherwise.

A subsequence of a string is a new string that is formed from the original string by deleting some (can be none) of the characters without disturbing the relative positions of the remaining characters. (i.e., "ace" is a subsequence of "abcde" while "aec" is not).

 

Example 1:

Input: s = "abc", t = "ahbgdc"
Output: true
Example 2:

Input: s = "axc", t = "ahbgdc"
Output: false
 

Constraints:

0 <= s.length <= 100
0 <= t.length <= 104
s and t consist only of lowercase English letters.


Solution:

Checked for edge cases, i.e if s is empty, then it is a subsequence of t.

Created a pointer then looped through t,
if the length of pointer is less than s and the char at that point is less equal to the pointer pointing at s (at the same index)
then the pointer is incremented by one within the loop then it repeats for the length of the string t.

if when the length of the pointer is equal to length of the subsequence string s, then it returns True

Outside, the for loop, it will return false if none of the if statement conditions are met. 


Time Complexity = O(n)
Space Complexity= O(1)