Create Single Number III - Leetcode
Problem:

Given an integer array nums, in which exactly two elements appear only once and all the other elements appear exactly twice. Find the two elements that appear only once. You can return the answer in any order.

Solution:

Initialised a check set. 
Looped through the nums array and if a number was in the check set already, it was removed. If it was not in the check set, then it will be added in the else statement.

Then at the end of, and outside the loop, the check set is converted to a list and returned  

Space Complexity: O(n)
Time Complexity: O(n)